+++ tempest_fastapi_sdk/__init__.py
@@ -372,7 +372,7 @@ from tempest_fastapi_sdk.websockets import (
     make_websocket_router,
 )
 
-__version__: str = "0.101.0"
+__version__: str = "0.102.0"
 
 __all__: list[str] = [
     "BASE_COLUMN_ORDER",